requirements.txt
Django==4.0.1
mysqlclient==2.1.0
Pillow==9.0.0
ipython==8.0.0
django-extensions==3.1.5
django-debug-toolbar==3.2.4
django-bootstrap4==21.2